Fix Create Agent event handling
Previously, the database manager was not updating the end block num of
old agents in the database. This commit makes sure that it complies
with the rules for slowly changing dimensions type 2 databases.

This would be an issue if there was ever a transaction added that
updated an agent object in state.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>